test3

+++ src/main/java/com/mindhub/homebanking/servicies/service/CardService.java
@@ -16,7 +16,7 @@ import java.time.LocalDate;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
+
 
 @Service
 public class CardService {
@@ -31,7 +31,7 @@ public class CardService {
         Map<String, Object> response = new HashMap<>();
         Client client = clientRepository.findByEmail(userMail);
         long cardNumber = RandomUtil.generateNumber(16);
-        int cardCVV = (int)RandomUtil.generateNumber(3);
+        long cardCVV = RandomUtil.generateNumber(3);
         Long cardCount = cardRepository.countByClientId(client.getId());
         if (cardCount >= 3 ) {
             response.put("success", false);
@@ -51,7 +51,7 @@ public class CardService {
             return response;
         }
 
-        Card newCard = new Card(CardType.valueOf(cardForm.cardType()), CardColor.valueOf(cardForm.cardColor()), cardNumber, cardCVV, LocalDate.now(), LocalDate.now().plusYears(5));
+        Card newCard = new Card(CardType.valueOf(cardForm.cardType()), CardColor.valueOf(cardForm.cardColor()), cardNumber, (int) cardCVV, LocalDate.now(), LocalDate.now().plusYears(5));
 
         newCard.setClient(client);
         newCard.setCardHolder(client);
